Full set of 24*4 body muscles implemented
Muscle cells enumeration according to
http://www.wormatlas.org/hermaphrodite/musclesomatic/Images/musfig15leg.htm
starting from DR quadrant. Enjoy!
Note: in real worm VL quadrant contains only 23 muscles (not 24).  In
simulation there are 24 muscle cells in this quadrant for symmetry.
They are completely symmetrical to those in VR quadrant, so send signals
for 22 and 23 real muscles to 23 and 24 muscles in simulation,
correspondingly (and nothing to 22-th muscle of simulation) for VL
quadrant.
Contact me if you have any questions.